I want to merge this change because it reduces the execution time of the query, when requesting `totalCount` with `stockAvailability` availability.

Setup used during tests:
- 400 Warehouses
- 557k Stocks
- 25k ProductVariants

Query used for testing:
```graphql
query products($channel: String!, $filter: ProductFilterInput, $first: Int, $after: String, $before: String, $last: Int, $sortBy: ProductOrder, $search: String) {
  products(
    channel: $channel
    filter: $filter
    first: $first
    after: $after
    before: $before
    last: $last
    sortBy: $sortBy
    search: $search
  ) {
    __typename
    edges {
      node{
        id
      }
      
    }
    pageInfo {
      __typename
      hasNextPage
      hasPreviousPage
      startCursor
      endCursor
    }
    totalCount
  }
}

```
Variables:
```json
{
  "channel": "default-channel",
  "first": 20,
  "filter": {
    "stockAvailability": "IN_STOCK"
  }
}
```
Results:
- 3.20:
  -  IN_STOCK:
      -  With `totalCount` in query:  `5.993674s`
      -  Without `totalCount` in query:   `0.322461s`
  -  OUT_OF_STOCK:
      -  With `totalCount` in query:  `6.600026s`
      -  Without `totalCount` in query:   `1.139934s`
- With the improvements:
  -  IN_STOCK:
      -  With `totalCount` in query:  `3.837313s`
      -  Without `totalCount` in query:   `0.307673s`
  -  OUT_OF_STOCK:
      -  With `totalCount` in query:  `3.966663s`
      -  Without `totalCount` in query:   `0.915774s`


Port of changes from: #16904



<!-- Please mention all relevant issue numbers. -->

# Impact

- [ ] New migrations
- [ ] New/Updated API fields or mutations
- [ ] Deprecated API fields or mutations
- [ ] Removed API types, fields, or mutations

# Docs

<!-- Docs are stored in a separate repository: https://github.com/saleor/saleor-docs/. -->
<!-- Please provide a link to the PR that updates documentation for your changes. -->
<!-- If changes in docs are not required, please mention that in the description. -->

- [ ] Link to documentation:

# Pull Request Checklist

<!-- Please keep this section. It will make the maintainer's life easier. -->

- [ ] Privileged queries and mutations are either absent or guarded by proper permission checks
- [ ] Database queries are optimized and the number of queries is constant
- [ ] Database migrations are either absent or optimized for zero downtime
- [ ] The changes are covered by test cases
- [ ] All new fields/inputs/mutations have proper labels added (`ADDED_IN_X`, `PREVIEW_FEATURE`, etc.)
- [ ] All migrations have proper dependencies
- [ ] All indexes are added concurrently in migrations
- [ ] All RunSql and RunPython migrations have revert option defined